+++ src/ir/QuantumComputation.cpp
@@ -22,7 +22,6 @@
 #include "ir/operations/SymbolicOperation.hpp"
 
 #include <algorithm>
-#include <array>
 #include <cassert>
 #include <cmath>
 #include <cstddef>
@@ -37,7 +36,6 @@
 #include <numeric>
 #include <optional>
 #include <ostream>
-#include <random>
 #include <ranges>
 #include <set>
 #include <sstream>
@@ -589,7 +587,7 @@ bool QuantumComputation::operator==(const QuantumComputation& rhs) const {
       initialLayout != rhs.initialLayout ||
       outputPermutation != rhs.outputPermutation ||
       ancillary != rhs.ancillary || garbage != rhs.garbage ||
-      seed != rhs.seed || globalPhase != rhs.globalPhase ||
+      globalPhase != rhs.globalPhase ||
       occurringVariables != rhs.occurringVariables) {
     return false;
   }
@@ -970,25 +968,12 @@ void QuantumComputation::checkClassicalRegister(
 void QuantumComputation::reverse() { std::ranges::reverse(ops); }
 
 QuantumComputation::QuantumComputation(const std::size_t nq,
-                                       const std::size_t nc,
-                                       const std::size_t s)
-    : seed(s) {
+                                       const std::size_t nc) {
   if (nq > 0) {
     addQubitRegister(nq);
   }
   if (nc > 0) {
     addClassicalRegister(nc);
-  }
-  if (seed != 0) {
-    mt.seed(seed);
-  } else {
-    // create and properly seed rng
-    std::array<std::mt19937_64::result_type, std::mt19937_64::state_size>
-        randomData{};
-    std::random_device rd;
-    std::ranges::generate(randomData, [&rd]() { return rd(); });
-    std::seed_seq seeds(std::begin(randomData), std::end(randomData));
-    mt.seed(seeds);
   }
 }
 
@@ -997,8 +982,8 @@ QuantumComputation::QuantumComputation(const QuantumComputation& qc)
       name(qc.name), quantumRegisters(qc.quantumRegisters),
       classicalRegisters(qc.classicalRegisters),
       ancillaRegisters(qc.ancillaRegisters), ancillary(qc.ancillary),
-      garbage(qc.garbage), mt(qc.mt), seed(qc.seed),
-      globalPhase(qc.globalPhase), occurringVariables(qc.occurringVariables),
+      garbage(qc.garbage), globalPhase(qc.globalPhase),
+      occurringVariables(qc.occurringVariables),
       initialLayout(qc.initialLayout), outputPermutation(qc.outputPermutation) {
   ops.reserve(qc.ops.size());
   for (const auto& op : qc.ops) {
@@ -1015,8 +1000,6 @@ QuantumComputation::operator=(const QuantumComputation& qc) {
     quantumRegisters = qc.quantumRegisters;
     classicalRegisters = qc.classicalRegisters;
     ancillaRegisters = qc.ancillaRegisters;
-    mt = qc.mt;
-    seed = qc.seed;
     globalPhase = qc.globalPhase;
     occurringVariables = qc.occurringVariables;
     initialLayout = qc.initialLayout;